uuid in path + batching
